backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pydantic-settings
black
pytest
httpx
python-jose
sqlalchemy
pytest-mock
google-auth[requests]
psycopg2-binary
pydantic[email]
python-multipart
PyPDF2
